+++ @navikt/core/react/src/alert/global-alert/root/GlobalAlertRoot.tsx
